Project file structure changed to modular
The project's file structure is module. There is a core module
and a ui module with corresponding pom files, along with a
root pom for the entire project .gitignore files have been
updated to reflect this, as well as module-info files.
At current commit, the project only boots with vscode and
does not run with maven.